tests/unit/com/opengamma/financial/covariance/ExponentialWeightedMovingAverageHistoricalVolatilityCalculatorTest.java
/**
 * Copyright (C) 2009 - 2009 by OpenGamma Inc.
 * 
 * Please see distribution for license.
 */
package com.opengamma.financial.covariance;

import static org.junit.Assert.assertEquals;

import org.junit.Test;

import com.opengamma.util.timeseries.DoubleTimeSeries;

/**
 * 
 * @author emcleod
 */
public class ExponentialWeightedMovingAverageHistoricalVolatilityCalculatorTest extends HistoricalVolatilityCalculatorTestCase {
  private static final HistoricalVolatilityCalculator<DoubleTimeSeries<Long>> CALCULATOR = new ExponentialWeightedMovingAverageHistoricalVolatilityCalculator<DoubleTimeSeries<Long>>(
      0.94, RETURN_CALCULATOR);

  @SuppressWarnings("unchecked")
  @Test
  public void test() {
    assertEquals(Math.sqrt(252) * CALCULATOR.evaluate(CLOSE_TS), 0.2455, EPS);
  }

  /*
   * (non-Javadoc)
   * 
   * @see
   * com.opengamma.financial.covariance.HistoricalVolatilityCalculatorTestCase
   * #getCalculator()
   */
  @Override
  protected HistoricalVolatilityCalculator<DoubleTimeSeries<Long>> getCalculator() {
    return CALCULATOR;
  }
}
